pyproject.toml
[tool.poetry]
name = "wanted"
version = "0.1.0"
description = ""
authors = ["sgaeng <ksge1124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
django-cors-headers = "^4.3.0"
pillow = "^10.0.1"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
